Added positional updating for strafing and driving forward
Also added scaling method to Vector to use diagonal strafe distance scaled by sqrt(2) for diagonal strafe positional updating

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Hardware/MRGyro.java
@@ -43,8 +43,8 @@ public class MRGyro {
             return heading;
     }
 
-    public static int convertToDegrees(double radians) {
-        return (int) (radians * 180 / Math.PI);
+    public static double convertToDegrees(double radians) {
+        return (radians * 180 / Math.PI);
     }
-    public static double convertToRadians(int degrees) { return degrees * Math.PI / 180; }
+    public static double convertToRadians(double degrees) { return degrees * Math.PI / 180; }
 }
